Consolidate CLI for webapp and workers - Move csw svc secrets to cs secrets - Set up dev mode for webapp that uses:   - DEBUG: true   - Devserver for code reloading.   - Maps local code directory to pod code directory.
