+++ Dockerfile.dev
@@ -2,4 +2,4 @@ FROM webbase
 
 RUN conda install -c conda-forge pylint black --yes
 
-CMD gunicorn --bind 0.0.0.0:$PORT webapp.wsgi
+CMD python manage.py runserver 0.0.0.0:$PORT
